FEAT: mapping member to board and comment

+++ server/fly-away/src/main/java/com/codestates/flyaway/domain/comment/service/CommentService.java
@@ -1,10 +1,11 @@
 package com.codestates.flyaway.domain.comment.service;
 
 import com.codestates.flyaway.domain.board.entity.Board;
-import com.codestates.flyaway.domain.board.repository.BoardRepository;
 import com.codestates.flyaway.domain.board.service.BoardService;
 import com.codestates.flyaway.domain.comment.entity.Comment;
 import com.codestates.flyaway.domain.comment.repository.CommentRepository;
+import com.codestates.flyaway.domain.member.entity.Member;
+import com.codestates.flyaway.domain.member.service.MemberService;
 import com.codestates.flyaway.global.exception.BusinessLogicException;
 import com.codestates.flyaway.global.exception.ExceptionCode;
 import com.codestates.flyaway.web.comment.dto.CommentDto;
@@ -16,6 +17,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
+
+import static com.codestates.flyaway.global.exception.ExceptionCode.NOT_AUTHORIZED;
 import static com.codestates.flyaway.web.comment.dto.CommentDto.CommentResponseDto.*;
 
 @RequiredArgsConstructor
@@ -26,13 +30,15 @@ public class CommentService {
 
     private final CommentRepository commentRepository;
     private final BoardService boardService;
+    private final MemberService memberService;
 
     public CommentResponseDto write(CommentDto.Write writeDto) {
 
+        Member member = memberService.findById(writeDto.getMemberId());
         Board board = boardService.findById(writeDto.getBoardId());
         board.addCommentCount();
-
         Comment comment = new Comment(writeDto.getContent());
+        comment.setMember(member);
         comment.setBoard(board);
         commentRepository.save(comment);
 
@@ -42,6 +48,9 @@ public class CommentService {
     public CommentResponseDto update(CommentDto.Update updateDto) {
 
         final Comment comment = commentRepository.getReferenceById(updateDto.getCommentId());
+        if(!Objects.equals(comment.getId(), updateDto.getMemberId())) {
+            throw new BusinessLogicException(ExceptionCode.NOT_AUTHORIZED);
+        }
         comment.update(updateDto.getCommentId(), updateDto.getContent());
 
         return toResponseDto(comment);
@@ -53,9 +62,12 @@ public class CommentService {
 
         return commentRepository.findByBoardId(boardId, pageable);
     }
-    public void delete(Long commentId) {
+    public void delete(CommentDto.Delete deleteDto) {
 
-        Comment comment = findById(commentId);
+        Comment comment = findById(deleteDto.getCommentId());
+        if(!Objects.equals(comment.getMember().getId(), deleteDto.getMemberId())) {
+            throw new BusinessLogicException(NOT_AUTHORIZED);
+        }
         commentRepository.delete(comment);
     }
